Remove AutoIT dependency
Utilizes the Requests library to make smart check against the head data
to compare URLs. This removes the dependency of AutoIT, and allows the
program to be cleanly run from command-line without dedicating a machine
to it. The speed is now directly related to network response speed, and
is overall much faster.